Using Admin step as a parent class
